Use shouldRenderComponent to reduce rendering and recalculation
By creating a component that only takes relevant properties that would
cause a recalculation, we're free to redefine shouldRenderComponent to
compare object equality of all of the props. If any props change, render
(and recalculate); otherwise, no need to render or recalculate.  This
is similar to memoization, but using `react` lifecycle.

In order for this approach to work, you have to create a component that
_doesn't_ take an unrelated prop (i.e. `currentTheme`).